Added initial CasperJS tests, these will test the basic functionality of the app from a user perspective.
Currently the test to check wether a sound has been played on the host page fails, despite the sound being playable.  To test this functionality you need to open a host page, then in a new tab open a client page where you can choose which sound to play, then go back to the original host page to hear the sound.

PhantomJS doesn't have a (straightforward) way to keep a page open in the background while opening a new page.  When returning to the host page, it's unaware of anything that's happened in the past, so does not have any record of a sound having been played.